+++ playground/Lambda/Lambda.AppHost/Program.cs
@@ -36,6 +36,7 @@ builder.AddAWSAPIGatewayEmulator("APIGatewayEmulator", Aspire.Hosting.AWS.Lambda
 builder.AddAWSLambdaFunction<Projects.SQSProcessorFunction>("SQSProcessorFunction", "SQSProcessorFunction::SQSProcessorFunction.Function::FunctionHandler")
         .WithReference(awsSdkConfig)
         .WithSQSEventSource(sqsDemoQueue1)
+        .WithSQSEventSource(sqsDemoQueue2)
         // These references are not necessary. It is added to confirm duplicate
         // CDK output parameters are not attempted to be added.
         .WithReference(sqsDemoQueue1)