+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Release 2026-04-29
+
+### Aspire.Hosting.AWS (13.0.1)
+* Fix duplicate resource name when adding more than one SQS Event Source to a resource
+
 ## Release 2026-02-25
 
 ### Aspire.Hosting.AWS (13.0.0)